sample-4/userAuth/serializers.py
from pyexpat import model
from rest_framework import serializers
from userAuth.models import  CustomUser
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework_simplejwt.tokens import RefreshToken, TokenError

class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('id' , 'password', 'first_name', 'last_name', 'email', 'date_of_birth', 'age', 'is_superuser', 'is_staff', 'phone_number', 'address', )


class LogoutSerializer(serializers.Serializer):
    refresh = serializers.CharField()

    def validate(self, attrs):
        self.token = attrs['refresh']
        return attrs

    def save(self, **kwargs):
        try:
            RefreshToken(self.token).blacklist()
        except TokenError:
            self.fail('bad token')



