sample-4/requirements.txt
django==3.2
djangorestframework
psycopg2
djangorestframework-simplejwt
django-environ